ScrollBoundaryContainer: use generated docs (multiple exports)
